Remove potion buckets
